Move server binaries to /cmd directory
Also moves /server/main.go to /cmd/internal/serverutil, because this part
of the "server" package is fairly specific to those binaries.
