+++ cmd/internal/serverutil/main.go
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package server holds code for running Trillian servers.
-package server
+// Package serverutil holds code for running Trillian servers.
+package serverutil
 
 import (
 	"context"
