+++ integration/functions.sh
@@ -118,8 +118,8 @@ log_prep_test() {
   local log_signer_count=${2:-1}
 
   echo "Building Trillian log code"
-  go build ${GOFLAGS} github.com/google/trillian/server/trillian_log_server/
-  go build ${GOFLAGS} github.com/google/trillian/server/trillian_log_signer/
+  go build ${GOFLAGS} github.com/google/trillian/cmd/trillian_log_server/
+  go build ${GOFLAGS} github.com/google/trillian/cmd/trillian_log_signer/
 
   # Wipe the test database
   yes | bash "${TRILLIAN_PATH}/scripts/resetdb.sh"
@@ -281,7 +281,7 @@ map_prep_test() {
   local rpc_server_count=${1:-1}
 
   echo "Building Trillian map code"
-  go build ${GOFLAGS} github.com/google/trillian/server/trillian_map_server/
+  go build ${GOFLAGS} github.com/google/trillian/cmd/trillian_map_server/
 
   # Wipe the test database
   yes | bash "${TRILLIAN_PATH}/scripts/resetdb.sh"
